% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adintel_to_sql.R
\name{adintel_to_sql}
\alias{adintel_to_sql}
\title{Replace "adintel" data types with "(postgre)SQL" data types}
\usage{
adintel_to_sql(adintel_type, r_out = FALSE)
}
\arguments{
\item{adintel_type}{a character vector.}

\item{check}{logical. Default is TRUE.}
}
\value{
a character vector.
}
\description{
Replace "adintel" data types with "(postgre)SQL" data types
}
\examples{
adintel_type <- c("Tinyint", "Char", "c")
adintel_to_sql(adintel_type)
}
