% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_match.R
\name{all_match}
\alias{all_match}
\title{Value Matching}
\usage{
all_match(x, table, which = TRUE)
}
\arguments{
\item{x}{vector or NULL: the values to be matched.}

\item{table}{vector or NULL: the values to be matched against.}

\item{which}{logical. If TRUE (default) the function returns the position of the elements of x without a match in the name of the output value.}
}
\value{
logical.
}
\description{
Value Matching
}
