% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numint.R
\name{as_numint}
\alias{as_numint}
\title{Numeric to integer or double}
\usage{
as_numint(x, warn = TRUE)
}
\arguments{
\item{x}{numeric vector.}

\item{warn}{logical.}
}
\value{
numeric or integer vector.
}
\description{
Numeric to integer or double
}
\examples{
x_noint <- c(2.1, 3)
x_int <- c(2, 3)
as_numint(x_noint)
as_numint(x_noint, warn = FALSE)
as_numint(x_int)
}
