% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coerce_numeric.R
\name{check_coerce_numeric}
\alias{check_coerce_numeric}
\alias{is_integer2}
\alias{is_double2}
\alias{is_numeric2}
\alias{as_numeric2}
\title{Check and coerce an object to a numeric type}
\usage{
is_integer2(x, n = NULL)

is_double2(x, n = NULL)

is_numeric2(x)

as_numeric2(x)
}
\arguments{
\item{x}{vector to be coerced.}

\item{n}{expected length of vector.}
}
\value{
is_xxx2 checks if an object (including a character vector) is of type xxx.
The function as_numeric2 automatically returns an integer if possible and a double otherwise..
}
\description{
Check and coerce an object to a numeric type
}
