% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_utf8.R
\name{check_utf8}
\alias{check_utf8}
\title{Check for the elements of a character object can be translated into valid UTF-8}
\usage{
check_utf8(x)
}
\arguments{
\item{x}{a character vector or a data frame (tibble, data.table...). If data frame the function gets applied only to character columns.}
}
\value{
a string reporting 'OK' if all elements can be transmated into valid UTF-8 or a message otherwise.
}
\description{
Check for the elements of a character object can be translated into valid UTF-8
}
\examples{

x <- c("fa\u00E7ile", "fa\xE7ile", "fa\xC3\xA7ile")
Encoding(x) <- c("UTF-8", "UTF-8", "bytes")
check_utf8(x)
check_utf8(data.frame(x = x, y = x))
}
