% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_to_num.R
\name{chr_to_num}
\alias{chr_to_num}
\title{Character vector to Numeric or Integert}
\usage{
chr_to_num(x, to_int = TRUE)
}
\arguments{
\item{x}{a character vector.}
}
\value{
logical.
}
\description{
Character vector to Numeric or Integert
}
\examples{
x_nonum <- c("a", "3", "4")
x_num <- c("2", "3", "4.1")
x_int <- c("2", "3", "4")
chr_to_num(x_nonum)
chr_to_num(x_num)
chr_to_num(x_int)
}
