% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_helpers.R
\name{date_helpers}
\alias{date_helpers}
\alias{is_year}
\alias{as_year}
\alias{is_month}
\alias{as_month}
\alias{is_day}
\alias{as_day}
\title{Helpers for dates}
\usage{
is_year(x)

as_year(x)

is_month(x)

as_month(x)

is_day(x)

as_day(x)
}
\arguments{
\item{x}{numeric or character.}

\item{y, m, d, sep}{characters.}

\item{as_chr}{logical.}
}
\value{
is_xxx returns TRUE or FALSE, as_xxx transforms input in the desired character format.
}
\description{
Helpers for dates
}
