% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{date_reference}
\alias{date_reference}
\title{Lookup Table for Dates (2009 - 2030)}
\format{
\subsection{\code{date_reference}}{

A data frame (tibble) with 152 rows and 3 columns.
\describe{
\item{media_type_id}{integer/numeric. The id for each specific media.}
\item{ukey_pos}{Type: integer/numeric. Positions of the columns forming the unique key.}
\item{ukey_man}{Type: character. Names (as per the AdIntel manual) of the columns forming the unique key for a specific 'media_type_id'.}
\item{media_type_id}{integer/numeric. The id for each specific media.}
}
}
}
\usage{
date_reference
}
\description{
Lookup Table for Dates (2009 - 2030)
}
\keyword{datasets}
