% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_file.R
\name{decompose_file}
\alias{decompose_file}
\title{Get info from file name}
\usage{
decompose_file(file, out_tibble = TRUE, out_unlist = FALSE)
}
\arguments{
\item{file}{string. File patth.}

\item{out_tibble}{logical. Default is TRUE. If TRUE the function returns a tibble. If FALSE, a list or a named character vector depending on 'out_unlist'.}

\item{out_unlist}{logical. Default is FALSE. If FALSE the function returns a list. If FALSE, a named character vector.}
}
\value{
a tibble, a list, or a named character vector with the follosing components: file, file_type_std, file_name_std (character), and adintel_year (integer),
}
\description{
Get info from file name
}
