% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding_utf8.R
\name{encoding_utf8}
\alias{encoding_utf8}
\title{Make sure econding is 'UTF-8'}
\usage{
encoding_utf8(x, from = "latin1", to = "UTF-8", sub = "*")
}
\arguments{
\item{x}{a character vector.}

\item{from}{a string.}

\item{to}{a string.}

\item{sub}{a string.}
}
\value{
the input character vector encoded as 'UTF-8'.
}
\description{
Make sure econding is 'UTF-8'
}
