% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fall_date.R
\name{fall_date}
\alias{fall_date}
\title{Date for Fall basaed on September eqwuinox}
\usage{
fall_date(year, yearweek = TRUE)
}
\arguments{
\item{year}{integer vector.}

\item{yearweek}{logical. If TRUE (default), the date reflects the Monday of the week of the Equinox. If False, the date of the equinox.}

\item{out_tibble}{logical. If TRUE (default), the function returns a tibble. A list otherwise.}
}
\value{
A tibble or a list.
}
\description{
Date for Fall basaed on September eqwuinox
}
