% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_to_info.R
\name{file_to_info}
\alias{file_to_info}
\title{Get info from file name}
\usage{
file_to_info(file)
}
\arguments{
\item{.dyn_data_file}{character vector.}
}
\value{
tibble.
}
\description{
Get info from file name
}
