% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread_adintel.R
\name{fread_adintel}
\alias{fread_adintel}
\title{Function of 'fread' with corrections for adintel datasets}
\usage{
fread_adintel(file, col_names_std, col_classes, ...)
}
\arguments{
\item{file}{string.}

\item{col_names_std}{character vector.}

\item{col_classes}{character vector.}

\item{...}{for future development.}
}
\value{
a data.table.
}
\description{
Function of 'fread' with corrections for adintel datasets
}
