% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adintel_tables.R
\name{get_adintel_tables}
\alias{get_adintel_tables}
\title{Get Adintel Tables}
\usage{
get_adintel_tables(adintel_dir)
}
\arguments{
\item{adintel_dir}{a string}
}
\value{
a tibble
}
\description{
Get Adintel Tables
}
