% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dynamic_file.R
\name{get_dynamic_file}
\alias{get_dynamic_file}
\title{Get all dynamic files paths}
\usage{
get_dynamic_file(dir, year, type, file)
}
\arguments{
\item{dir}{string.}

\item{year}{integer}

\item{type}{string.}

\item{file}{string.}
}
\value{
string.
}
\description{
Get all dynamic files paths
}
