% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_static_file.R
\name{get_static_file}
\alias{get_static_file}
\title{Get files for static references}
\usage{
get_static_file(dir, year, file = NULL, dir_out = FALSE)
}
\arguments{
\item{dir}{string.}

\item{year}{integer}

\item{file}{string.}

\item{dir_out}{logical. Default is FALSE}
}
\value{
string.
}
\description{
Get files for static references
}
