% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adintel_tbl_class.R
\name{indintel_tbl_class}
\alias{indintel_tbl_class}
\alias{new_adintel_tbl_cols}
\alias{new_adintel_range_partition}
\alias{new_adintel_list_partition}
\alias{new_adintel_tbl_indices}
\title{Class adintel_tbl}
\usage{
new_adintel_tbl_cols(
  x = list(),
  col_names_std = character(),
  col_classes = character(),
  sql_datatypes = character()
)

new_adintel_range_partition(
  x = list(),
  partition_by = character(),
  partition_name = character(),
  range = list(from = numeric(), to = numeric())
)

new_adintel_list_partition(
  x = list(),
  partition_by = character(),
  partition_name = character(),
  values
)

new_adintel_tbl_indices(
  x = list(),
  index_cols = character(),
  pk_cols = character(),
  fk_tbl = data.frame()
)
}
\description{
Class adintel_tbl
}
