% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_lower.R
\name{is_lower}
\alias{is_lower}
\title{Checks for Lower Case Characters}
\usage{
is_lower(x, which = FALSE)
}
\arguments{
\item{x}{a string.}

\item{which}{logical. Defauls is FALSE}
}
\value{
a vector. When which is FALSE it returns a vectors with TRUE/FALSE. When which is TRUE it returns the position of lowercase characters.
}
\description{
Checks for Lower Case Characters
}
