% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_numint.R
\name{is_numint}
\alias{is_numint}
\title{Check if a numeric vector is integer}
\usage{
is_numint(x)
}
\arguments{
\item{x}{numeric vvctor.}
}
\value{
logical.
}
\description{
Check if a numeric vector is integer
}
\examples{
x_noint <- c(2.1, 3)
x_int <- c(2, 3)
is_numint(x_noint)
is_numint(x_int)

}
