% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_upper.R
\name{is_upper}
\alias{is_upper}
\title{Checks for Upper Case Characters}
\usage{
is_upper(x, which = FALSE)
}
\arguments{
\item{x}{a string.}

\item{which}{logical. Defauls is FALSE}
}
\value{
a vector. When which is FALSE it returns a vectors with TRUE/FALSE. When which is TRUE it returns the position of uppercase characters.
}
\description{
Checks for Upper Case Characters
}
