% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup_date}
\alias{lookup_date}
\title{Lookup Table for Dates (2009 - 2030)}
\format{
\subsection{\code{lookup_date}}{

A data frame (tibble) with 152 rows and 3 columns.
\describe{
\item{date}{Date.}
\item{year}{integer/numeric Year.}
\item{month}{integer/numeric Month.}
\item{mday}{integer/numeric Day of the month.}
\item{yquarter}{character. Ready for tsibble::yearquarter.}
\item{ymonth}{character. Ready for tsibble::yearmonth.}
\item{yweek}{character. Ready for tsibble::yearweek.}
\item{is_53weeks}{logical. TRUE if yweek has 53 weeks.}
\item{wday}{integer/numeric. Day of the week. Monday = 1.}
\item{wday_lbl}{integer/numeric. Day of the week as character.}
\item{is_leap}{logical. TRUE if leap year.}
}
}
}
\usage{
lookup_date
}
\description{
Lookup Table for Dates (2009 - 2030)
}
\keyword{datasets}
