% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grep_cmd.R
\name{make_grep_cmd}
\alias{make_grep_cmd}
\title{Create cmd}
\usage{
make_grep_cmd(file, ..., first_n = NULL, opts = "-wE")
}
\arguments{
\item{file}{a string.}

\item{...}{.}

\item{first_n}{an integer.}

\item{opts}{a string.}
}
\value{
a vector.
}
\description{
Create cmd
}
