% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_helpers.R, R/sql_make_code.R
\name{not_helpers}
\alias{not_helpers}
\alias{not_in}
\alias{\%notin\%}
\alias{not_na}
\alias{not_null}
\alias{not_empty}
\alias{sql_make_fields}
\alias{sql_create_tbl}
\alias{sql_create_part_tbl}
\title{CHelpers for "not" selections}
\usage{
not_in(x, table)

x \%notin\% table

not_na(x)

not_null(x)

not_empty(x)

sql_make_fields(.con, .tbl_name, .col_names, .data_types, .pk = NA)

sql_create_tbl(
  .con,
  .tbl_name,
  .col_names,
  .data_types,
  .pk = NULL,
  .year,
  .date_from,
  .date_to
)

sql_create_part_tbl(.part_type, .part_name, .tbl_name, .part_values)
}
\arguments{
\item{x}{vector.}

\item{table}{vector.}

\item{.con}{DBIconnection.}

\item{.tbl_name}{character.}

\item{.col_names}{character.}

\item{.data_types}{character.}

\item{.pk}{character.}
}
\value{
vector.
}
\description{
CHelpers for "not" selections

SQL codes
}
