% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric padding.R
\name{numeric_padding}
\alias{numeric_padding}
\alias{numpad}
\alias{numpad2}
\alias{numpad4}
\title{Numeric values as characters with '0' padding}
\usage{
numpad(x, width = NULL)

numpad2(x)

numpad4(x)
}
\arguments{
\item{x}{numeric or character coercible to numeric.}

\item{width}{numeric.}
}
\value{
a character
}
\description{
Numeric values as characters with '0' padding
}
