% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numpad.R, R/string_helpers.R
\name{numpad}
\alias{numpad}
\alias{numpad2}
\alias{numpad4}
\title{Numeric values as chracters with '0' 0padding}
\usage{
numpad(x, width = 0L)

numpad2(x)

numpad4(x)

numpad(x, width = 0L)
}
\arguments{
\item{x}{numeric. Alternatively, character coercible to integer.}

\item{width}{numeric.}
}
\value{
a character
}
\description{
Numeric values as chracters with '0' 0padding

Title
}
