% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occurrences_categories}
\alias{occurrences_categories}
\title{Nielsen Adintel Dataset - Media Categories}
\format{
\subsection{\code{occurrences_categories}}{

A data frame (tibble) with 24 rows and 5 columns.
\describe{
\item{media_type_id}{integer/numeric. The id for each specific media.}
\item{media_type}{Type: character. An informative name for each 'media_type_id'. It is divided in two parts by the separator "__" (two underscores). Left of the separator is the name of the media (always one string), Right of the separator are abbreviation for the characteristics of the media separated by one underscore. E.g. 'network_tv__nat_spa' indicates that the 'media_type_id' refers to data for Network TV at the local level ("loc") and Spanish language ("spa").}
\item{media_category}{Type: character. c("tv", "print", "radio", "outdoor", "online", "cinema"). }
\item{media_geo}{Type: character. It indicates wheter a certain media is 'national' or 'local'.}
\item{is_spanish}{Type: logical. TRUE if the 'media_type_id' is in Spanish language.}
}
}
}
\usage{
occurrences_categories
}
\description{
This table provides a lookup table for each of the media in the AdIntel dataset.
}
\keyword{datasets}
