% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nrows.R
\name{read_nrows}
\alias{read_nrows}
\alias{read_nrows_1}
\title{Read the number of rows of a file}
\usage{
read_nrows_1(x, header)

read_nrows(file, header = TRUE)
}
\arguments{
\item{header}{logical.}

\item{file}{a character vector.}
}
\value{
a list or a tibble/data.frame.
}
\description{
Read the number of rows of a file
}
