% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nrows_i.R
\name{read_nrows_i}
\alias{read_nrows_i}
\title{Read the number of rows of a file}
\usage{
read_nrows_i(file, header = TRUE, list_out = FALSE, df_out = FALSE)
}
\arguments{
\item{file}{a string.}

\item{header}{logical.}

\item{list_out}{logical.}

\item{df_out}{logical.}
}
\value{
a list or a tibble/data.frame.
}
\description{
Read the number of rows of a file
}
