% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_adintel.R
\name{rename_adintel}
\alias{rename_adintel}
\title{Rename Adintel Variables}
\usage{
rename_adintel(x, named = FALSE)
}
\arguments{
\item{x}{character vector.}
}
\value{
character vector.
}
\description{
Rename Adintel Variables
}
