% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_adintel_cols.R
\name{rename_adintel_cols}
\alias{rename_adintel_cols}
\title{Rename Adintel Variables}
\usage{
rename_adintel_cols(x, named = F)
}
\arguments{
\item{name_manual}{character vector.}
}
\value{
character vector.
}
\description{
Rename Adintel Variables
}
