% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safread.R
\name{safread_warn}
\alias{safread_warn}
\title{Safe fread}
\usage{
safread_warn(
  input = "",
  file = NULL,
  text = NULL,
  cmd = NULL,
  sep = "auto",
  sep2 = "auto",
  dec = ".",
  quote = "\\"",
  nrows = Inf,
  header = "auto",
  na.strings = getOption("datatable.na.strings", "NA"),
  stringsAsFactors = FALSE,
  verbose = getOption("datatable.verbose", FALSE),
  skip = "__auto__",
  select = NULL,
  drop = NULL,
  colClasses = NULL,
  integer64 = getOption("datatable.integer64", "integer64"),
  col.names,
  check.names = FALSE,
  encoding = "unknown",
  strip.white = TRUE,
  fill = FALSE,
  blank.lines.skip = FALSE,
  key = NULL,
  index = NULL,
  showProgress = getOption("datatable.showProgress", interactive()),
  data.table = getOption("datatable.fread.datatable", TRUE),
  nThread = getDTthreads(verbose),
  logical01 = getOption("datatable.logical01", FALSE),
  keepLeadingZeros = getOption("datatable.keepLeadingZeros", FALSE),
  yaml = FALSE,
  autostart = NA,
  tmpdir = tempdir(),
  tz = "UTC"
)
}
\description{
Safe fread
}
