% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sed_replace.R
\name{sed_replace}
\alias{sed_replace}
\alias{sed_replace_all}
\title{Replace string (gsub) in file using 'sed'}
\usage{
sed_replace(
  file_in,
  file_out = NULL,
  pattern,
  replacement,
  all = FALSE,
  tmp_pattern = "sed_temp_"
)

sed_replace_all(file_in, file_out = NULL, pattern, replacement)
}
\arguments{
\item{file_in}{character.}

\item{file_out}{string.}

\item{pattern}{string.}

\item{replacement}{string.}

\item{all}{logical.}

\item{tmp_pattern}{string.}
}
\value{
a list indicating: input file, output file, pattern: replacement, exit code.
}
\description{
Replace string (gsub) in file using 'sed'
}
