% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_slice.R
\name{slice_vec}
\alias{slice_vec}
\title{Vector Slicing}
\usage{
slice_vec(x, i)
}
\arguments{
\item{x}{vector.}

\item{i}{integer or integer vector. Negative integers indicate positions from the end.}
}
\value{
a vector of the same kind of x.
}
\description{
Vector Slicing
}
