% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_to_info.R, R/sql_helpers.R
\name{sql_helpers}
\alias{sql_helpers}
\alias{sql_tbl_name}
\alias{sql_define_datatypes}
\alias{adintel_to_sql}
\alias{sql_define_cols}
\alias{sql_create_tbl}
\title{Get info from file name}
\usage{
sql_tbl_name(.file_type_std, .file_name_std)

sql_define_datatypes(.list_manual, .str_manual = NULL, output)

adintel_to_sql(adintel_type, r_out = FALSE)

sql_define_cols(.col_name, .col_type, .n1 = NA, .n2 = NA, ...)

sql_create_tbl(tbl_name, col_def)
}
\arguments{
\item{.col_name}{character.}

\item{.col_type}{character.}

\item{.n1}{integer. Defaults to NA.}

\item{.n2}{integer. Defaults to NA.}

\item{file}{character vector.}
}
\value{
glue string.

glue string.
}
\description{
Get info from file name

SQL column definition block
}
