% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_make_coldef.R
\name{sql_make_coldef}
\alias{sql_make_coldef}
\title{SQL column definition block}
\usage{
sql_make_coldef(.name, .type, .n1 = NA, .n2 = NA, .constraints = NA, ...)
}
\arguments{
\item{.name}{character.}

\item{.type}{character.}

\item{.n1}{integer. Defaults to NA.}

\item{.n2}{integer. Defaults to NA.}

\item{.constraints}{character. Defaults to NA.}
}
\value{
glue string.
}
\description{
SQL column definition block
}
