% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_make_code.R
\name{sql_tbl_creator}
\alias{sql_tbl_creator}
\alias{sql_make_fields}
\alias{sql_build_tbl}
\alias{sql_part_by_range}
\alias{sql_part_by_list}
\alias{sql_create_part}
\title{SQL codes}
\usage{
sql_make_fields(.tbl_name, .col_names, .data_types, .pk = NA)

sql_build_tbl(.tbl_name, .col_names, .data_types, .pk = NA, .part_col, ...)

sql_part_by_range(.part_col, ....)

sql_part_by_list(.part_col, ....)

sql_create_part(
  .tbl_name,
  .part_name,
  .from = NULL,
  .to = NULL,
  .values = NULL,
  ...
)
}
\arguments{
\item{.tbl_name}{character.}

\item{.col_names}{character.}

\item{.data_types}{character.}

\item{.pk}{character.}

\item{.con}{DBIconnection.}
}
\value{
SQL.
}
\description{
SQL codes
}
