% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_type_to_code.R
\name{sql_type_to_code}
\alias{sql_type_to_code}
\title{Make sql code for data type}
\usage{
sql_type_to_code(var_type, precision_var, scale_var, check = T)
}
\arguments{
\item{var_type}{vector}

\item{precision_var}{vector}

\item{scale_var}{vector}

\item{check}{logical}
}
\value{
vector
}
\description{
Make sql code for data type
}
