% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_end_date.R
\name{start_end_date}
\alias{start_end_date}
\title{Start and End Date in a file}
\usage{
start_end_date(year, file)
}
\arguments{
\item{year}{integer.}

\item{file}{character}
}
\value{
dates.
}
\description{
Start and End Date in a file
}
