% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_sep_upper.R
\name{str_sep_upper}
\alias{str_sep_upper}
\title{Separates String Based on Uppercase Characters}
\usage{
str_sep_upper(x_chr, sep = " ", to_lower = TRUE, to_list = FALSE, named = TRUE)
}
\arguments{
\item{x_chr}{a string or a character vector of the form "AaBBCccDd".}

\item{sep}{a string indicating the separator. Default is a space " ".}

\item{to_lower}{logical. Default TRUE.}

\item{to_list}{logical. Default FALSE.}

\item{named}{logical. Default TRUE.}
}
\value{
a string or a character vector of the form "Aa BB Ccc Dd" (or Aa_BB_Ccc_Dd").
}
\description{
Separates String Based on Uppercase Characters
}
