% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helpers.R
\name{string_helpers}
\alias{string_helpers}
\alias{no_case}
\alias{is_upper}
\alias{is_lower}
\alias{str_upper_uscore}
\alias{rename_adintel}
\title{String helpers}
\usage{
no_case(pattern)

is_upper(x, which = FALSE)

is_lower(x, which = FALSE)

str_upper_uscore(x, to_lower = TRUE)

rename_adintel(x, named = TRUE)
}
\arguments{
\item{pattern}{character.}
}
\value{
a character
}
\description{
String helpers
}
