% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data_type.R
\name{test_data_type}
\alias{test_data_type}
\title{Test data type of tables vs manual}
\usage{
test_data_type(x_data, x_manual, check = FALSE)
}
\arguments{
\item{x_data}{vector.}

\item{x_manual}{vector.}

\item{check}{logical.}
}
\value{
character vector.
}
\description{
Test data type of tables vs manual
}
