% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{timer}
\alias{timer}
\title{Timer}
\usage{
timer(start, n_pre = 0, n_post = 0, msg = "Time elapsed: {.x}")
}
\arguments{
\item{start}{Sys.time object0}

\item{n_pre}{integer.}

\item{n_post}{integer.}

\item{msg}{character.}
}
\value{
character
}
\description{
Timer
}
\examples{
t0 <- Sys.time()
Sys.sleep(5)
timer(start = t0, n_pre = 1, n_post = 2, msg = "Time elapsed: {.x}")
}
