% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_c.R
\name{to_c}
\alias{to_c}
\title{To character vector}
\usage{
to_c(...)
}
\arguments{
\item{...}{unquoted names or one string of names separated by ','.}
}
\value{
a character vector.
}
\description{
To character vector
}
\examples{
to_c("a,b,c")
}
