% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sql_datatype.R
\name{to_sql_datatype}
\alias{to_sql_datatype}
\title{Create sql name - datatype}
\usage{
to_sql_datatype(name, type, precision = NA, scale = NA, adintel = TRUE)
}
\arguments{
\item{name}{character vector.}

\item{type}{character vector.}

\item{precision}{integer vector.}

\item{scale}{integer vector.}

\item{adintel}{logical. Default is TRUE.}
}
\value{
a character vector.
}
\description{
Create sql name - datatype
}
