% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unquoted_to_chr.R
\name{unquoted_to_chr}
\alias{unquoted_to_chr}
\title{Turns unquoted names to strings.}
\usage{
unquoted_to_chr(...)
}
\arguments{
\item{...}{unquoted names.}
}
\value{
a character vector.
}
\description{
Turns unquoted names to strings.
}
