% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_db.R
\name{write_to_db}
\alias{write_to_db}
\alias{adintel_read_tsv}
\alias{write_to_db_con}
\title{Title}
\usage{
adintel_read_tsv(tbl_name, file, col_names, col_classes = NULL, col_uk = NULL)

write_to_db(
  con,
  tbl_name,
  file,
  year,
  media_type_id = NA,
  col_names,
  col_classes,
  col_uk,
  overwrite,
  append
)

write_to_db_con(con)
}
\arguments{
\item{tbl_name}{string.}

\item{file}{string.}

\item{col_names}{character vector.}

\item{col_classes}{character vector.}

\item{col_uk}{character vector.}

\item{con}{a db connection.}

\item{year}{integer.}

\item{overwrite}{logical.}

\item{append}{logical.}
}
\value{
nothing. Must be used (write_to_db_con) with purrr::pwalk
}
\description{
Title
}
