% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_to_info.R
\name{file_to_info}
\alias{file_to_info}
\alias{file_to_info_dynamic}
\alias{file_to_info_static}
\title{Get info from file name}
\usage{
file_to_info_dynamic(file)

file_to_info_static(file)
}
\arguments{
\item{file}{character vector.}
}
\value{
glue string.
}
\description{
Get info from file name
}
