% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_helpers.R
\name{not_helpers}
\alias{not_helpers}
\alias{not_in}
\alias{\%notin\%}
\alias{not_na}
\alias{not_null}
\alias{not_empty}
\title{CHelpers for "not" selections}
\usage{
not_in(x, table)

x \%notin\% table

not_na(x)

not_null(x)

not_empty(x)
}
\arguments{
\item{x}{vector.}

\item{table}{vector.}
}
\value{
vector.
}
\description{
CHelpers for "not" selections
}
