% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_helpers.R, R/sql_make_code.R
\name{not_helpers}
\alias{not_helpers}
\alias{not_in}
\alias{\%notin\%}
\alias{not_na}
\alias{not_null}
\alias{not_empty}
\alias{sql_make_fields}
\alias{sql_create_tbl}
\title{CHelpers for "not" selections}
\usage{
not_in(x, table)

x \%notin\% table

not_na(x)

not_null(x)

not_empty(x)

sql_make_fields(.tbl_name, .col_names, .data_types, .pk = NA)

sql_create_tbl(.tbl_name, .col_names, .data_types, .pk = NA)
}
\arguments{
\item{x}{vector.}

\item{table}{vector.}

\item{.tbl_name}{character.}

\item{.col_names}{character.}

\item{.data_types}{character.}

\item{.pk}{character.}

\item{.con}{DBIconnection.}
}
\value{
vector.
}
\description{
CHelpers for "not" selections

SQL codes
}
