% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrences_dates_from_to.R
\name{occurrences_dates_from_to}
\alias{occurrences_dates_from_to}
\alias{occ_date_range}
\title{Starting and Ending Dates of Occurrences Files}
\usage{
occ_date_range(.file_name_std, year, date_to = FALSE)
}
\arguments{
\item{date_to}{logical. FALSE is equivalent to 'occ_date_from' and TRUE to 'occ_date_to}

\item{occ_file}{character vector.}
}
\value{
vector of dates as characters.
}
\description{
Starting and Ending Dates of Occurrences Files
}
