% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{read_files}
\alias{read_files}
\alias{read_nrows}
\title{Read the number of rows of a file}
\usage{
read_nrows(file)
}
\arguments{
\item{file}{a character vector.}

\item{header}{logical.}
}
\value{
a list or a tibble/data.frame.
}
\description{
Read the number of rows of a file
}
