% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nrows.R
\name{read_nrows}
\alias{read_nrows}
\title{Read the number of rows of a file}
\usage{
read_nrows(full_name, df_out = TRUE)
}
\arguments{
\item{full_name}{a string.}

\item{df_out}{logical. If TRUE (default) it returns a tibble (or data.frame).}
}
\value{
a list or a tibble/data.frame.
}
\description{
Read the number of rows of a file
}
