% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nrows.R
\name{read_nrows}
\alias{read_nrows}
\title{Read the number of rows of a file}
\usage{
read_nrows(full_name, list_out = FALSE, df_out = FALSE)
}
\arguments{
\item{full_name}{a string.}

\item{df_out}{logical.}
}
\value{
a list or a tibble/data.frame.
}
\description{
Read the number of rows of a file
}
