% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_helpers.R
\name{sql_helpers}
\alias{sql_helpers}
\alias{sql_define_datatypes}
\alias{adintel_to_sql}
\alias{sql_define_cols}
\alias{sql_create_tbl}
\title{SQL column definition block}
\usage{
sql_define_datatypes(.list_manual, .str_manual = NULL, output)

adintel_to_sql(adintel_type, r_out = FALSE)

sql_define_cols(.col_name, .col_type, .n1 = NA, .n2 = NA, ...)

sql_create_tbl(tbl_name, col_def)
}
\arguments{
\item{.col_name}{character.}

\item{.col_type}{character.}

\item{.n1}{integer. Defaults to NA.}

\item{.n2}{integer. Defaults to NA.}
}
\value{
glue string.
}
\description{
SQL column definition block
}
