% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helpers.R
\name{string_helpers}
\alias{string_helpers}
\alias{no_case}
\title{String helpers}
\usage{
no_case(pattern)
}
\arguments{
\item{pattern}{character.}
}
\value{
a character
}
\description{
String helpers
}
