% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helpers.R
\name{string_helpers}
\alias{string_helpers}
\alias{no_case}
\alias{is_upper}
\alias{is_lower}
\alias{str_upper_uscore}
\alias{rename_adintel}
\alias{str_split_comma}
\alias{str_adintel_to_title}
\alias{str_embrace}
\alias{str_embrace_sq}
\alias{str_embrace_cur}
\title{String helpers}
\usage{
no_case(pattern)

is_upper(x, which = FALSE)

is_lower(x, which = FALSE)

str_upper_uscore(x, to_lower = TRUE)

rename_adintel(x, named = TRUE)

str_split_comma(x)

str_adintel_to_title(x)

str_embrace(x, left = "(", right = ")")

str_embrace_sq(x, left = "[", right = "]")

str_embrace_cur(x, left = "{", right = "}")
}
\arguments{
\item{pattern}{a string The pattern for which to ignore case.}

\item{x}{a character vector.}

\item{which}{logical. It returns indices instead of the actual characters.}

\item{to_lower}{logical. If TRUE (default) the output is all lower case.}

\item{named}{logical. If TRUE, the output vector has the original strings as names.}
}
\value{
Same type as input.
}
\description{
String helpers
}
