% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table_to_db.R
\name{write_to_db}
\alias{write_to_db}
\alias{uk_unique_fn}
\alias{adintel_read_tsv}
\title{Title}
\usage{
uk_unique_fn(.df, .uk)

adintel_read_tsv(
  tbl_name,
  file,
  col_names,
  col_classes = NULL,
  col_uk = NULL,
  out_df_only = FALSE
)

write_to_db(con, tbl_name, df, overwrite, append)
}
\arguments{
\item{tbl_name}{string.}

\item{file}{string.}

\item{col_names}{character vector.}

\item{col_classes}{character vector.}

\item{col_uk}{character vector.}

\item{con}{a db connection.}

\item{overwrite}{logical.}

\item{append}{logical.}

\item{year}{integer.}
}
\value{
nothing. Must be used (write_to_db_con) with purrr::pwalk
}
\description{
Title
}
